Create potential national number from the copy of normalized national number
The function maybeStripNationalPrefixAndCarrierCode will modify the underlying data of its first argument (the national number).
With the current implementation, the value of normalizedNationalNumber could be altered unintentionally by this function, as
potentialNationalNumber and normalizedNationalNumber are sharing the same underlying byte slice. To solve the problem, we can
create potentialNationalNumber from a copy of normalizedNationalNumber.